Moved season-hub and its components to app directory
