+++ app/components/GrandPrixInfoHeader/index.tsx
