+++ app/layout.tsx
@@ -12,7 +12,7 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <body>
+      <body className="bg-gray-50 dark:bg-gray-900">
         <Navbar />
         {children}
         <Footer />
