+++ app/page.tsx
@@ -1,11 +1,10 @@
-import { NextPage } from "next";
 import Hero from "./components/Hero";
 import Image from "next/image";
 import Features from "./components/Features";
 import Faq from "./components/Faq";
 import telemetryDemo from "../public/images/readme/telemetry-demo.png";
 
-const Home: NextPage = () => {
+const Home = () => {
   return (
     <>
       <Hero />
